refactor: Repository의 DTO 네이밍 수정
