+++ backend/src/main/java/org/donggle/backend/application/repository/dto/MemberInfo.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.application.repository;
+package org.donggle.backend.application.repository.dto;
 
 public record MemberInfo(
         Long id,
